refactor: into a component
