+++ .projenrc.ts
@@ -1,7 +1,6 @@
-import { readdirSync } from 'fs';
-import path from 'path';
-import { ReleasableCommits, SourceCode, awscdk, github, javascript, release } from 'projen';
+import { ReleasableCommits, awscdk, github, javascript, release } from 'projen';
 import { ArrowParens, NodePackageManager } from 'projen/lib/javascript';
+import { SubPathExports } from './projenrc/sub-path-exports';
 
 let cdkVersion = '2.168.0';
 const project = new awscdk.AwsCdkConstructLibrary({
@@ -108,31 +107,6 @@ project.addTask('integ:update', {
   receiveArgs: true,
 });
 
-const solutions = readdirSync(path.join(project.outdir, project.srcdir), {
-  encoding: 'utf8',
-  withFileTypes: true,
-})
-  .filter(entry => entry.isDirectory())
-  .map(entry => entry.name)
-  .sort((a, b) => a.localeCompare(b));
-
-const sourceCode = new SourceCode(project, path.join(project.srcdir, 'index.ts'));
-sourceCode.line('// ' + sourceCode.marker);
-
-const subPathExportPath = (...chunks: string[]) => './' + path.posix.join(project.libdir, ...chunks);
-
-const subPathExports: Record<string, string> = {
-  '.': subPathExportPath('index.js'),
-};
-
-for (const solution of solutions) {
-  const exportName = solution.split('-').join('_');
-
-  sourceCode.line(`export * as ${exportName} from './${solution}';`);
-
-  subPathExports[`./${solution}`] = subPathExportPath(solution, 'index.js');
-}
-
-project.package.addField('exports', subPathExports);
+new SubPathExports(project);
 
 project.synth();
